refactor: extract swap non alloyed and rebalancer pass
